% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{generateComplete}
\alias{generateComplete}
\title{generateComplete}
\usage{
generateComplete(outcomeType, numSamp, beta0, beta1, beta2, xmean, xsd, errorsd)
}
\arguments{
\item{outcomeType}{Is the outcome variable binary or continuous?}

\item{numSamp}{Number of observations in the sample}

\item{beta0}{Intercept}

\item{beta1}{Treatment effect}

\item{beta2}{Slope of auxiliary variable}

\item{xmean}{Mean of auxiliary variable}

\item{xsd}{Standard deviation of auxiliary variable}

\item{errorsd}{Parameter to add noise. Standard deviation of errors}
}
\description{
This function generate complete datasets.
}
