% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{generateData}
\alias{generateData}
\title{generateData}
\usage{
generateData(mechanism, outcomeType, propMiss = 0.1, numSamp = 1000,
  beta0 = 1, beta1 = 2, beta2 = 1, xmean = 0, xsd = 1,
  errorsd = 0.1, rr = 5)
}
\arguments{
\item{mechanism}{What is the missing data mechanism?}

\item{outcomeType}{Is the outcome variable binary or continuous?}

\item{propMiss}{Expected proportion of missingness. Defaults to 0.1.}

\item{numSamp}{Number of observations in the sample. Defaults to 1000.}

\item{beta0}{Intercept. Defaults to 1.}

\item{beta1}{Treatment effect. Defaults to 2.}

\item{beta2}{Slope of auxiliary variable. Defaults to 1.}

\item{xmean}{Mean of auxiliary variable. Defaults to 0.}

\item{xsd}{Standard deviation of auxiliary variable. Defaults to 1.}

\item{errorsd}{Standard deviation of error term. Defaults to 1.}

\item{rr}{Risk ratio. Defaults to 5. Defaults to 5.}
}
\description{
This function creates a complete dataset and then generates missingness
}
