% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{makeMissing}
\alias{makeMissing}
\title{makeMissing}
\usage{
makeMissing(outcomeType, mechanism, numSamp, propMiss, rr, df, beta0, beta1,
  beta2, xmean)
}
\arguments{
\item{outcomeType}{Is the outcome variable binary or continuous?}

\item{mechanism}{What is the missing data mechanism?}

\item{numSamp}{Number of observations in the sample}

\item{propMiss}{Expected proportion of missingness}

\item{rr}{Risk ratio}

\item{df}{The complete dataframe}

\item{beta0}{Intercept}

\item{beta1}{Treatment effect}

\item{beta2}{Slope of auxiliary variable}

\item{xmean}{Mean of auxiliary variable}
}
\description{
This function take a complete dataset and generates missingness
}
