% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{runCC}
\alias{runCC}
\title{runCC}
\usage{
runCC(df, outcomeType)
}
\arguments{
\item{df}{The dataset with missing values generated by generateData()}

\item{outcomeType}{Is the outcome variable binary or continuous?}
}
\description{
This function creates a models using complete case analysis
}
