% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{runImputation}
\alias{runImputation}
\title{runImputation}
\usage{
runImputation(df, m = 5, method, interaction = FALSE, donors = 3,
  outcomeType)
}
\arguments{
\item{df}{The dataset with missing values generated by generateData()}

\item{m}{Number of imputations.}

\item{method}{The method of imputing univariate missingness}

\item{interaction}{Want to add interaction to the imputation model? Defaults to FALSE.}

\item{donors}{Number of donors if using predictive mean matching. Defaults to 3.}

\item{outcomeType}{Is the outcome variable binary or continuous?}
}
\description{
This function creates a models via multiple imputation and returns model parameters
}
