% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{runSimulation}
\alias{runSimulation}
\title{#' runSimulation}
\usage{
runSimulation(numSim = 1000, m = 5, mechanism = "mcar", propMiss = 0.1,
  numSamp = 1000, beta0 = 1, beta1 = 2, beta2 = 1, xmean = 0,
  xsd = 1, errorsd = 0.5, rr = 5, outcomeType)
}
\arguments{
\item{numSim}{Number of simulations. Defaults to 1000.}

\item{m}{Number of imputations. Defaults to 5.}

\item{mechanism}{What is the missing data mechanism? Defaults to mcar.}

\item{propMiss}{Expected proportion of missingness. Defaults to 0.1.}

\item{numSamp}{Number of observations in the sample. Defaults to 1000.}

\item{beta0}{Intercept. Defaults to 1.}

\item{beta1}{Treatment effect. Defaults to 2.}

\item{beta2}{Slope of auxiliary variable. Defaults to 1.}

\item{xmean}{Mean of auxiliary variable. Defaults to 0.}

\item{xsd}{Standard deviation of auxiliary variable. Defaults to 1.}

\item{errorsd}{Standard deviation of error term. Defaults to 0.5.}

\item{rr}{Risk ratio. Defaults to 5. Defaults to 5.}

\item{outcomeType}{Is the outcome variable binary or continuous?}
}
\description{
This function runs a simulations fitting models using various methods of handling missing data. Passes back an array of model parameter values.
}
