% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rctmiss_functions.R
\name{runTrue}
\alias{runTrue}
\title{runTrue}
\usage{
runTrue(df, outcomeType)
}
\arguments{
\item{df}{The dataset with missing values generated by generateData()}

\item{outcomeType}{Is the outcome variable binary or continuous?}
}
\description{
This function creates a models using the true values of y
}
